#261: Split Storage
