+++ core/loadModules.js
@@ -4,6 +4,7 @@
 
 /* Dependencies */
 import Notifications from '../lib/Notifications.js';
+import Request from '../lib/Request.js';
 import Uploads from '../lib/Uploads.js';
 import User from '../lib/User.js';
 
@@ -15,6 +16,7 @@ import User from '../lib/User.js';
  */
 export default async function loadModules(next) {
 	this.user = new User(this);
+	this.request = new Request(this);
 
 	if (this.config.mail) {
 		this.notifications = new Notifications(this);
