+++ lib/Storage.js
@@ -12,7 +12,6 @@ import moment from 'moment';
 import { console } from './Cluster.js';
 import SaplingError from './SaplingError.js';
 import Response from './Response.js';
-import Validation from './Validation.js';
 
 
 /* Default user structure */
@@ -92,71 +91,6 @@ export default class Storage {
 
 
 	/**
-	 * Take an incoming request and make sense of it
-	 *
-	 * @param {object} request The request object from Express
-	 */
-	parseRequest(request) {
-		/* Get the URL segments from the requested URL */
-		const query = new URL(request.url, `${request.protocol}://${request.hostname}`);
-		const parts = query.pathname.split('/');
-
-		/* Request method */
-		const method = request.method && request.method.toUpperCase();
-
-		/* Trim uneeded parts of the request */
-		if (parts[0] === '') {
-			parts.splice(0, 1);
-		}
-
-		if (parts[parts.length - 1] === '') {
-			parts.splice(-1, 1);
-		}
-
-		if (parts[0] === 'data') {
-			parts.splice(0, 1);
-		}
-
-		/* Name each of the URL segments */
-		const collection = parts[0];
-		const fields = [];
-		const values = [];
-
-		for (let i = 1; i < parts.length; i += 2) {
-			fields.push(parts[i]);
-			values.push(parts[i + 1]);
-		}
-
-		console.log('Request', collection, fields, values);
-
-		/* Leave a warning if no permission on a writable request */
-		if ((method === 'POST' || method === 'DELETE') && !request.permission) {
-			console.warn(`You should add a permission for \`${request.url}\`.`);
-		}
-
-		/* Convert URLSearchParams to a regular object */
-		const queryObject = {};
-		for (const queryKey of query.searchParams.keys()) {
-			if (query.searchParams.getAll(queryKey).length > 1) {
-				queryObject[queryKey] = query.searchParams.getAll(queryKey);
-			} else {
-				queryObject[queryKey] = query.searchParams.get(queryKey);
-			}
-		}
-
-		/* Modify the request object */
-		return _.extend(request, {
-			collection,
-			fields,
-			values,
-			query: queryObject, // Query params
-			type: parts.length >= 3 ? 'filter' : 'all',
-			isLogged: Boolean(request.session && request.session.user),
-		});
-	}
-
-
-	/**
 	 * Import DB driver if needed; no-op if already loaded.
 	 */
 	async importDriver() {
@@ -224,276 +158,6 @@ export default class Storage {
 
 
 	/**
-	 * Returns an array of fields that should
-	 * be omitted from the response due to permissions.
-	 *
-	 * @param {string} role The role being checked
-	 * @param {string} collection The collection being checked against
-	 */
-	disallowedFields(role, collection) {
-		/* Get the collection definition */
-		const rules = this.schema[collection];
-		const omit = [];
-
-		/* Loop every field in the collection */
-		for (const key in rules) {
-			if (Object.prototype.hasOwnProperty.call(rules, key)) {
-				const rule = rules[key];
-
-				/* Normalise the access rule to be an object with r,w */
-				const access = typeof rule.access === 'string' ? {
-					r: rule.access,
-					w: rule.access,
-				} : rule.access;
-
-				/* Skip if not defined or anyone can view */
-				if (!access || access.r === 'anyone' || access.r === 'owner') {
-					continue;
-				}
-
-				/* Leave out the fields that the viewer can't access */
-				if (this.app.user.isRoleAllowed(role, access.r) === false) {
-					omit.push(key);
-				}
-			}
-		}
-
-		return omit;
-	}
-
-
-	/**
-	 * Get a list of fields in a given collection that only "owner"
-	 * level users are allowed to see.
-	 *
-	 * @param {string} collection The collection being checked
-	 */
-	ownerFields(collection) {
-		/* Get the collection definition */
-		const rules = this.schema[collection];
-		const fields = [];
-
-		/* Loop every field in the collection */
-		for (const key in rules) {
-			if (Object.prototype.hasOwnProperty.call(rules, key)) {
-				const rule = rules[key];
-
-				/* Normalise the access rule to be an object with r */
-				const access = typeof rule.access === 'string' ? {
-					r: rule.access,
-				} : rule.access;
-
-				/* Get the fields that are owner-only */
-				if (access && access.r === 'owner') {
-					fields.push(key);
-				}
-			}
-		}
-
-		return fields;
-	}
-
-
-	/**
-	 * Validate the data of a given POST request
-	 *
-	 * @param {object} request Request object from Express
-	 * @param {object} response Response object from Express
-	 */
-	validateData(request, response) {
-		const { collection, body, session, type } = request;
-
-		/* Get the collection definition */
-		const rules = this.schema[collection] || {};
-
-		let errors = [];
-		const data = body || {};
-
-		/* Get the role from session, if any */
-		const role = this.getRole({ session });
-
-		/* Model must be defined before pushing data */
-		if (!rules && this.config.strict) {
-			new Response(this.app, request, response, new SaplingError({
-				status: '500',
-				code: '1010',
-				title: 'Non-existent',
-				detail: 'This model does not exist.',
-				meta: {
-					type: 'data',
-					error: 'nonexistent',
-				},
-			}));
-			return false;
-		}
-
-		/* Go through every key in incoming data */
-		for (const key in data) {
-			if (Object.prototype.hasOwnProperty.call(data, key)) {
-				/* Ignore CSRF tokens */
-				if (key === '_csrf') {
-					delete data[key];
-				}
-
-				/* Get the corresponding ruleset */
-				const rule = rules[key];
-
-				/* Trim incoming data unless otherwise specified in model */
-				if (typeof data[key] === 'string' && (!rule || !('trim' in rule) || rule.trim !== false)) {
-					data[key] = String(data[key]).trim();
-				}
-
-				/* If the field isn't defined */
-				if (!rule) {
-					/* In strict mode, don't allow unknown fields */
-					if (this.config.strict) {
-						console.warn('UNKNOWN FIELD', key);
-						delete data[key];
-					}
-
-					/* Otherwise skip this field */
-					continue;
-				}
-
-				const dataType = (rule.type || rule).toLowerCase();
-
-				/* If the data is a number, convert from string */
-				if (dataType === 'number') {
-					data[key] = Number.parseFloat(data[key], 10);
-				}
-
-				/* Test in the validation library */
-				const error = new Validation().validate(data[key], key, rule);
-				if (error.length > 0) {
-					errors = error;
-				}
-
-				/* If this field has no defined access level, we can skip the rest of the checks */
-				if (!rule.access) {
-					continue;
-				}
-
-				/* Get the write access level */
-				const access = rule.access.w || rule.access;
-
-				/* If the field is owner-only, defer to individual op methods to check against it */
-				if (access === 'owner') {
-					continue;
-				}
-
-				/* If we do not have access, raise hell */
-				if (this.app.user.isRoleAllowed(role, access) === false) {
-					console.warn(`NO ACCESS TO FIELD '${key}'`);
-					console.warn('Current permission level:', role);
-					console.warn('Required permission level:', access);
-					delete data[key];
-				}
-			}
-		}
-
-		/* For insertion, handle required fields and default values */
-		if (type === 'all') {
-			/* Go through every defined field */
-			for (const key in rules) {
-				/* If the value exists, it will have already been validated above */
-				if (key in data) {
-					continue;
-				}
-
-				if (typeof rules[key] !== 'object') {
-					continue;
-				}
-
-				/* We now know the given field does not have a corresponding value
-				   in the incoming data */
-
-				/* If required, generate error */
-				if (rules[key].required) {
-					errors.push({
-						status: '422',
-						code: '1001',
-						title: 'Invalid Input',
-						detail: `You must provide a value for key \`${key}\``,
-						meta: {
-							key,
-							rule: 'required',
-						},
-					});
-				}
-
-				/* Set the data to the default value, if provided */
-				if ('default' in rules[key]) {
-					data[key] = rules[key].default;
-				}
-			}
-		}
-
-		console.log('ERRORS', errors);
-
-		return errors.length > 0 ? errors : [];
-	}
-
-
-	/**
-	 * Parse URL segments into "where" clause constraints
-	 *
-	 * @param {object} request Request object from Express
-	 */
-	getConstraints(request) {
-		const conditions = {};
-
-		/* Add any and all constraints to the query */
-		if (request.type === 'filter' && request.fields.length === request.values.length) {
-			for (let index = 0; index < request.fields.length; ++index) {
-				let values = request.values[index].split(',');
-				if (values.length === 1) {
-					values = values[0];
-				}
-
-				conditions[request.fields[index]] = values;
-			}
-		}
-
-		return conditions;
-	}
-
-
-	/**
-	 * Add a constraint for the creator to match the currently logged in
-	 * user, if appropriate for the given request and role
-	 *
-	 * @param {object} request Request object from Express
-	 * @param {string} role User role
-	 */
-	getCreatorConstraint(request, role) {
-		const conditions = {};
-
-		if (request.permission && request.permission.role.includes('owner') && role !== 'admin') {
-			conditions._creator = request.session.user._id;
-		}
-
-		return conditions;
-	}
-
-
-	/**
-	 * Get the user role from the session
-	 *
-	 * @param {object} request Request object from Express
-	 * @returns {string/null} Role as string, or null if not logged in
-	 */
-	getRole(request) {
-		let role = null;
-
-		if (request.session && request.session.user) {
-			role = request.session.user.role;
-		}
-
-		return role;
-	}
-
-
-	/**
 	 * Serve an incoming GET request from the database
 	 *
 	 * @param {object} request Request object from Express
@@ -503,7 +167,7 @@ export default class Storage {
 		await this.importDriver();
 
 		/* Prepare the raw request */
-		request = this.parseRequest(request);
+		request = this.app.request.parse(request);
 
 		/* Get the collection definition */
 		const rules = this.schema[request.collection] || {};
@@ -511,7 +175,7 @@ export default class Storage {
 		let conditions = {};
 
 		/* Get the user role from session */
-		const role = this.getRole(request);
+		const role = this.app.user.getRole(request);
 
 		/* Check if we're logged in */
 		this.app.user.isUserAuthenticatedForRoute(request, response);
@@ -546,7 +210,7 @@ export default class Storage {
 
 		/* Parse "where" clause, if any */
 		/* If the permission is "owner", check that we are the owner, or an admin */
-		conditions = _.extend(conditions, this.getConstraints(request), this.getCreatorConstraint(request, role));
+		conditions = _.extend(conditions, this.app.request.getConstraints(request), this.app.request.getCreatorConstraint(request, role));
 
 		/* If we have reference fields */
 		const references = [];
@@ -573,10 +237,10 @@ export default class Storage {
 			const array = await this.db.read(request.collection, conditions, options, references);
 
 			/* Get the list of fields we should not be able to see */
-			const omit = this.disallowedFields(role, request.collection);
+			const omit = this.app.user.disallowedFields(role, this.schema[request.collection]);
 
 			/* Get the list of fields that only owners can see */
-			const ownerFields = this.ownerFields(request.collection);
+			const ownerFields = this.app.user.ownerFields(this.schema[request.collection]);
 
 			/* Process fields against both lists */
 			for (let i = 0; i < array.length; ++i) {
@@ -616,7 +280,7 @@ export default class Storage {
 		await this.importDriver();
 
 		/* Prepare the raw request */
-		request = this.parseRequest(request);
+		request = this.app.request.parse(request);
 
 		/* Check if we're logged in */
 		this.app.user.isUserAuthenticatedForRoute(request, response);
@@ -628,13 +292,13 @@ export default class Storage {
 		const data = request.body;
 
 		/* Validate the updated data */
-		const errors = this.validateData(request, response);
+		const errors = this.app.request.validateData(request, response);
 		if (errors.length > 0) {
 			return new Response(this.app, request, response, new SaplingError(errors));
 		}
 
 		/* Get the user role from session */
-		const role = this.getRole(request);
+		const role = this.app.user.getRole(request);
 
 		/* Specially format certain fields */
 		for (const field of Object.keys(rules)) {
@@ -683,7 +347,7 @@ export default class Storage {
 		if (request.type === 'filter') {
 			/* Parse "where" clause, if any */
 			/* If the permission is "owner", check that we are the owner, or an admin */
-			conditions = _.extend(conditions, this.getConstraints(request), this.getCreatorConstraint(request, role));
+			conditions = _.extend(conditions, this.app.request.getConstraints(request), this.app.request.getCreatorConstraint(request, role));
 
 			/* Update hidden fields */
 			data._lastUpdated = Date.now();
@@ -727,16 +391,16 @@ export default class Storage {
 		await this.importDriver();
 
 		/* Prepare the raw request */
-		request = this.parseRequest(request);
+		request = this.app.request.parse(request);
 
 		let conditions = {};
 
 		/* Get the user role from session */
-		const role = this.getRole(request);
+		const role = this.app.user.getRole(request);
 
 		/* Parse "where" clause, if any */
 		/* If the permission is "owner", check that we are the owner, or an admin */
-		conditions = _.extend(conditions, this.getConstraints(request), this.getCreatorConstraint(request, role));
+		conditions = _.extend(conditions, this.app.request.getConstraints(request), this.app.request.getCreatorConstraint(request, role));
 
 		/* Send it to the database */
 		return await this.db.remove(request.collection, conditions);
