+++ lib/User.js
@@ -1,7 +1,7 @@
 /**
  * User
  *
- * Built-in user account functionality
+ * Built-in user account and permissions functionality
  */
 
 /* Dependencies */
@@ -152,5 +152,89 @@ export default class User {
 
 		/* Default to anyone */
 		return ['anyone'];
+	}
+
+
+	/**
+	 * Get the user role from the session
+	 *
+	 * @param {object} request Request object from Express
+	 * @returns {string/null} Role as string, or null if not logged in
+	 */
+	getRole(request) {
+		let role = null;
+
+		if (request.session && request.session.user) {
+			role = request.session.user.role;
+		}
+
+		return role;
+	}
+
+
+	/**
+	 * Returns an array of fields that should
+	 * be omitted from the response due to permissions.
+	 *
+	 * @param {string} role The role being checked
+	 * @param {string} rules The rules of the collection being checked against
+	 */
+	disallowedFields(role, rules) {
+		const omit = [];
+
+		/* Loop every field in the collection */
+		for (const key in rules) {
+			if (Object.prototype.hasOwnProperty.call(rules, key)) {
+				const rule = rules[key];
+
+				/* Normalise the access rule to be an object with r,w */
+				const access = typeof rule.access === 'string' ? {
+					r: rule.access,
+					w: rule.access,
+				} : rule.access;
+
+				/* Skip if not defined or anyone can view */
+				if (!access || access.r === 'anyone' || access.r === 'owner') {
+					continue;
+				}
+
+				/* Leave out the fields that the viewer can't access */
+				if (this.isRoleAllowed(role, access.r) === false) {
+					omit.push(key);
+				}
+			}
+		}
+
+		return omit;
+	}
+
+
+	/**
+	 * Get a list of fields in a given model that only "owner"
+	 * level users are allowed to see.
+	 *
+	 * @param {string} rules The model being checked
+	 */
+	ownerFields(rules) {
+		const fields = [];
+
+		/* Loop every field in the collection */
+		for (const key in rules) {
+			if (Object.prototype.hasOwnProperty.call(rules, key)) {
+				const rule = rules[key];
+
+				/* Normalise the access rule to be an object with r */
+				const access = typeof rule.access === 'string' ? {
+					r: rule.access,
+				} : rule.access;
+
+				/* Get the fields that are owner-only */
+				if (access && access.r === 'owner') {
+					fields.push(key);
+				}
+			}
+		}
+
+		return fields;
 	}
 }
